+++ apps/desktop/src/routes/(window-chrome)/new-main/useSystemHardwareOptions.ts
@@ -1,0 +1,106 @@
+import { useQuery } from "@tanstack/solid-query";
+import {
+	listAudioDevices,
+	listScreens,
+	listVideoDevices,
+	listWindows,
+} from "~/utils/queries";
+import { useRecordingOptions } from "../OptionsContext";
+import type {
+	CameraInfo,
+	DeviceOrModelID,
+	ScreenCaptureTarget,
+} from "~/utils/tauri";
+
+const findCamera = (cameras: CameraInfo[], id: DeviceOrModelID) => {
+	return cameras.find((c) => {
+		if (!id) return false;
+		return "DeviceID" in id
+			? id.DeviceID === c.device_id
+			: id.ModelID === c.model_id;
+	});
+};
+
+export function useSystemHardwareOptions() {
+	const { rawOptions, setOptions } = useRecordingOptions();
+	const screens = useQuery(() => listScreens);
+	const windows = useQuery(() => listWindows);
+	const cameras = useQuery(() => listVideoDevices);
+	const mics = useQuery(() => listAudioDevices);
+
+	cameras.promise.then((cameras) => {
+		if (rawOptions.cameraID && findCamera(cameras, rawOptions.cameraID)) {
+			setOptions("cameraLabel", null);
+		}
+	});
+
+	mics.promise.then((mics) => {
+		if (rawOptions.micName && !mics.includes(rawOptions.micName)) {
+			setOptions("micName", null);
+		}
+	});
+
+	const options = {
+		screen: () => {
+			let screen;
+
+			if (rawOptions.captureTarget.variant === "display") {
+				const screenId = rawOptions.captureTarget.id;
+				screen =
+					screens.data?.find((s) => s.id === screenId) ?? screens.data?.[0];
+			} else if (rawOptions.captureTarget.variant === "area") {
+				const screenId = rawOptions.captureTarget.screen;
+				screen =
+					screens.data?.find((s) => s.id === screenId) ?? screens.data?.[0];
+			}
+
+			return screen;
+		},
+		window: () => {
+			let win;
+
+			if (rawOptions.captureTarget.variant === "window") {
+				const windowId = rawOptions.captureTarget.id;
+				win = windows.data?.find((s) => s.id === windowId) ?? windows.data?.[0];
+			}
+
+			return win;
+		},
+		camera: () => {
+			if (!rawOptions.cameraID) return undefined;
+			return findCamera(cameras.data || [], rawOptions.cameraID);
+		},
+		micName: () => mics.data?.find((name) => name === rawOptions.micName),
+		target: (): ScreenCaptureTarget | undefined => {
+			switch (rawOptions.captureTarget.variant) {
+				case "display": {
+					const screen = options.screen();
+					if (!screen) return;
+					return { variant: "display", id: screen.id };
+				}
+				case "window": {
+					const window = options.window();
+					if (!window) return;
+					return { variant: "window", id: window.id };
+				}
+				case "area": {
+					const screen = options.screen();
+					if (!screen) return;
+					return {
+						variant: "area",
+						bounds: rawOptions.captureTarget.bounds,
+						screen: screen.id,
+					};
+				}
+			}
+		},
+	};
+
+	return {
+		screens,
+		windows,
+		cameras,
+		mics,
+		options,
+	};
+}
